+++ scripts/json_parser.py
@@ -1,5 +1,5 @@
 import json
-from call_ai_function import call_ai_function
+from ai_functions import fix_json
 from config import Config
 cfg = Config()
 
@@ -47,30 +47,3 @@ def fix_and_parse_json(json_str: str, try_to_fix_with_gpt: bool = True):
               return json_str
           else:
             raise e
-        
-def fix_json(json_str: str, schema: str, debug=False) -> str:
-    # Try to fix the JSON using gpt:
-    function_string = "def fix_json(json_str: str, schema:str=None) -> str:"
-    args = [json_str, schema]
-    description_string = """Fixes the provided JSON string to make it parseable and fully complient with the provided schema.\n If an object or field specifed in the schema isn't contained within the correct JSON, it is ommited.\n This function is brilliant at guessing when the format is incorrect."""
-
-    # If it doesn't already start with a "`", add one:
-    if not json_str.startswith("`"):
-      json_str = "```json\n" + json_str + "\n```"
-    result_string = call_ai_function(
-        function_string, args, description_string, model=cfg.fast_llm_model
-    )
-    if debug:
-        print("------------ JSON FIX ATTEMPT ---------------")
-        print(f"Original JSON: {json_str}")
-        print("-----------")
-        print(f"Fixed JSON: {result_string}")
-        print("----------- END OF FIX ATTEMPT ----------------")
-    try:
-        return json.loads(result_string)
-    except:
-        # Get the call stack:
-        # import traceback
-        # call_stack = traceback.format_exc()
-        # print(f"Failed to fix JSON: '{json_str}' "+call_stack)
-        return "failed"
